--- src/main/java/techcourse/myblog/exception/exception/AuthenticationException.java
@@ -1,11 +1,0 @@
-package techcourse.myblog.exception.exception;
-
-public class AuthenticationException extends RuntimeException {
-    public AuthenticationException() {
-        super("권한이 없습니다.");
-    }
-
-    public AuthenticationException(String message) {
-        super(message);
-    }
-}
