+++ src/main/java/techcourse/myblog/service/LoginService.java
@@ -3,10 +3,10 @@ package techcourse.myblog.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import techcourse.myblog.domain.User;
-import techcourse.myblog.dto.UserRequestDto;
-import techcourse.myblog.dto.UserResponseDto;
 import techcourse.myblog.exception.LoginException;
 import techcourse.myblog.repository.UserRepository;
+import techcourse.myblog.service.dto.UserRequestDto;
+import techcourse.myblog.service.dto.UserResponseDto;
 import techcourse.myblog.utils.converter.UserConverter;
 
 @Service
