+++ src/main/java/techcourse/myblog/utils/converter/ArticleConverter.java
@@ -2,7 +2,7 @@ package techcourse.myblog.utils.converter;
 
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.User;
-import techcourse.myblog.dto.ArticleRequestDto;
+import techcourse.myblog.service.dto.ArticleRequestDto;
 
 public class ArticleConverter {
     public static Article toEntity(ArticleRequestDto articleRequestDto, User author) {
