+++ src/main/java/techcourse/myblog/web/argument/LoginUserArgumentResolver.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import techcourse.myblog.exception.exception.AuthenticationException;
-import techcourse.myblog.dto.UserResponseDto;
+import techcourse.myblog.exception.AuthenticationException;
+import techcourse.myblog.service.dto.UserResponseDto;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
