+++ src/test/java/techcourse/myblog/template/LoginTemplate.java
@@ -4,9 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.reactive.server.WebTestClient;
-import org.springframework.web.reactive.function.BodyInserters;
 import reactor.core.publisher.Mono;
-import techcourse.myblog.dto.UserRequestDto;
+import techcourse.myblog.service.dto.UserRequestDto;
 
 public class LoginTemplate extends SignUpTemplate {
     private static final Logger log = LoggerFactory.getLogger(LoginTemplate.class);
